+++ CHANGELOG.md
@@ -93,3 +93,4 @@
 - MVP-450 - update add company logo
 - MVP-460 - Create page types
 - MVP-512 - Handle user requesting generic content
+- MVP-496 - Handle signup from export plan